heaptest kinda broken and stufe also add just test mode

+++ build.sh
@@ -1,9 +1,12 @@
 #!/bin/bash
 
 SAVE_TEMPS=false
+JUST_TEST=false
 for arg in "$@"; do
 	if [ "$arg" = "-save-temps" ]; then
 		SAVE_TEMPS=true
+	elif [ "$arg" = "-t" ]; then
+		JUST_TEST=true
 	fi
 done
 
@@ -31,40 +34,42 @@ catza() {
 	printf '\0'
 }
 
-mkdir -p build
-mkdir -p build/bitmaps
+if [ "$JUST_TEST" = false ]; then
+	mkdir -p build
+	mkdir -p build/bitmaps
 
-fasm src/boot.asm build/boot.bin
-fasm src/kernel.asm build/kernel.sys
-catza assets/text/boot/logo.txt >> build/kernel.sys
-catza assets/text/boot/text.txt >> build/kernel.sys
-fasm src/command.asm build/command.sys
-fasm src/helloworld.asm build/helloworld.exe
-fasm src/graphix.asm build/graphix.exe
-fasm src/basic.asm build/basic.exe
-fasm src/heaptest.asm build/heaptest.exe
+	fasm src/boot.asm build/boot.bin
+	fasm src/kernel.asm build/kernel.sys
+	catza assets/text/boot/logo.txt >> build/kernel.sys
+	catza assets/text/boot/text.txt >> build/kernel.sys
+	fasm src/command.asm build/command.sys
+	fasm src/helloworld.asm build/helloworld.exe
+	fasm src/graphix.asm build/graphix.exe
+	fasm src/basic.asm build/basic.exe
+	fasm src/heaptest.asm build/heaptest.exe
 
-python3 tools/thinfs.py createbootable nsr-dos.img build/boot.bin NSRDOS
-add_to_disk nsr-dos.img \
-	build/kernel.sys \
-	build/command.sys \
-	build/helloworld.exe \
-	build/heaptest.exe \
-	build/basic.exe \
-	$(find docs/ -type f)
-truncate -s 1440k nsr-dos.img
+	python3 tools/thinfs.py createbootable nsr-dos.img build/boot.bin NSRDOS
+	add_to_disk nsr-dos.img \
+		build/kernel.sys \
+		build/command.sys \
+		build/helloworld.exe \
+		build/heaptest.exe \
+		build/basic.exe \
+		$(find docs/ -type f)
+	truncate -s 1440k nsr-dos.img
 
-cp assets/images/preconverted/* build/bitmaps
-convert_images
+	cp assets/images/preconverted/* build/bitmaps
+	convert_images
 
-python3 tools/thinfs.py create disk-2.img BDRIVE
-add_to_disk disk-2.img \
-	build/graphix.exe \
-	$(find build/bitmaps/ -type f)
-truncate -s 1440k disk-2.img
+	python3 tools/thinfs.py create disk-2.img BDRIVE
+	add_to_disk disk-2.img \
+		build/graphix.exe \
+		$(find build/bitmaps/ -type f)
+	truncate -s 1440k disk-2.img
 
-if [ "$SAVE_TEMPS" = false ]; then
-	rm -rf build/
+	if [ "$SAVE_TEMPS" = false ]; then
+		rm -rf build/
+	fi
 fi
 
 qemu-system-i386 -monitor stdio \
